Send BEGIN eagerly on transaction begin
